Added subdirectory support
By default, ostumake will look for OCaml code into every subdirectory
available.

It is possible to exclude files and/or directories by adding them to
the EXCL variable. The previous behavior is also available by setting
the NOSUBDIR variable.

Most of the code has been moved to the gnu.ocaml.obj.mk file. This
allows the user to choose between building a whole binary or just the
object files.

This commit closes #3.